Completed task 3

+++ 0x1A-hash_tables/3-hash_table_set.c
@@ -10,7 +10,7 @@
 
 int hash_table_set(hash_table_t *ht, const char *key, const char *value)
 {
-	int index;
+	unsigned long int index;
 	hash_node_t *node;
 
 	if (ht == NULL || key == NULL || strlen(key) == 0)
@@ -20,7 +20,7 @@ int hash_table_set(hash_table_t *ht, const char *key, const char *value)
 	node = malloc(sizeof(hash_node_t));
 	if (node == NULL)
 		return (0);
-
+	
 	node->key = strdup(key);
 	if (node->key == NULL)
 	{
